requirements.txt
earthengine_api==0.1.244
ee>=0.2
GDAL>=3.1.4
geopandas==0.9.0
imbalanced_learn==0.8.0
imblearn==0.0
matplotlib==3.4.1
numpy==1.19.4
#osgeo==0.0.1
osmnx==0.16.2
pandas==1.1.5
pyrosm>=0.6.0
PyYAML==6.0
rasterio==1.2.1
requests==2.25.0
scikit_image>=0.17.2
scikit_learn==1.1.0
Shapely==1.7.1
tqdm==4.62.3
sphinx>=4.3.0
sphinx_tabs>=3.2.0
sphinx-rtd-theme>=1.0.0
